pyproject.toml
[project]
name = "esa-ainstein-artifacts"
version = "0.1.0"
description = "Multi-Agent RAG System for Energy System Architecture"

readme = "README.md"
requires-python = ">=3.11,<3.14"

license = "Apache-2.0"
license-files = [
    "LICENSE"
]

authors = [
    { name = "Aion Project Contributors" }
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Vector DB / RAG
    "weaviate-client>=4.19.3",
    # Document ingestion
    "python-docx>=1.2.0",
    "pypdf>=6.9.1",
    "pymupdf>=1.27.1",
    "python-frontmatter>=1.1.0",
    # LLM / Agent
    "openai>=2.21.0",
    "pydantic-ai-slim[openai]>=0.2",
    # Async / IO
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    # CLI / UX
    "typer-slim==0.21.1",
    "python-dotenv>=1.2.1",
    # Config / models
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.0",
    # Web UI
    "fastapi>=0.129.0",
    "uvicorn>=0.34.0",
    # Observability
    "structlog>=24.4.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0,<9.0.3",  # 9.0.3+ adds runtime pygments import; pygments excluded by policy
    "pytest-asyncio>=1.3.0",
    "pytest-timeout>=2.3.1",
    "pytest-xdist>=3.5.0",
    "pytest-rerunfailures>=15.0",
    "black>=26.3.0",
    "ruff>=0.15.5",
]

[build-system]
requires = ["setuptools>=68.2.2", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
aion = "aion.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
markers = [
    "unit: Pure-logic tests, no external services needed (<30s total)",
    "functional: End-to-end tests requiring Weaviate + Ollama (~3-4 min)",
    "ingestion: Ingestion integrity checks against live Weaviate",
    "generation: ArchiMate generation pipeline tests",
    "benchmark: Retrieval quality benchmarks and provider comparison",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.uv]
package = true
override-dependencies = [
    "pygments ; python_version < '0'",  # Exclude: optional visual aid, CVE in 2.19.x
]
